# Flattened recursion