Pushing more performance out of route execution, removed recursive calls in favour of a loop.

Didn't really gain much performance...

Closes #11