@AndrewBurian 2017-04-26
Actually managed to shave about 6-14% off

before
```
BenchmarkSingleRoute-2      	 3000000	       486 ns/op	     208 B/op	       4 allocs/op
BenchmarkShallowAndWide-2   	 2000000	       885 ns/op	     222 B/op	       6 allocs/op
BenchmarkNarrowAndDeep-2    	  100000	     14475 ns/op	    4432 B/op	       6 allocs/op
BenchmarkFan-2              	 2000000	       854 ns/op	     214 B/op	       6 allocs/op
```

after
```
BenchmarkSingleRoute-2      	 3000000	       426 ns/op	     208 B/op	       4 allocs/op
BenchmarkShallowAndWide-2   	 2000000	       816 ns/op	     222 B/op	       6 allocs/op
BenchmarkNarrowAndDeep-2    	  100000	     13649 ns/op	    4432 B/op	       6 allocs/op
BenchmarkFan-2              	 2000000	       737 ns/op	     214 B/op	       6 allocs/op
```